simplify proto types
