+++ gamercade_interface/src/output/game.rs
@@ -13,13 +13,13 @@ pub struct UpdateGameRequest {
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GameInfoRequest {
+pub struct SingleGameRequest {
     #[prost(sfixed64, tag = "1")]
     pub game_id: i64,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct MultipleGamesInfoRequest {
+pub struct MultipleGamesRequest {
     #[prost(sfixed64, repeated, tag = "1")]
     pub game_ids: ::prost::alloc::vec::Vec<i64>,
 }
@@ -60,12 +60,6 @@ pub struct GameInfoDetailed {
 pub struct Author {
     #[prost(sfixed64, tag = "1")]
     pub user_id: i64,
-}
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct DeleteGameRequest {
-    #[prost(sfixed64, tag = "1")]
-    pub game_id: i64,
 }
 /// Generated client implementations.
 pub mod game_service_client {
@@ -154,7 +148,7 @@ pub mod game_service_client {
         }
         pub async fn get_single_game_info(
             &mut self,
-            request: impl tonic::IntoRequest<super::GameInfoRequest>,
+            request: impl tonic::IntoRequest<super::SingleGameRequest>,
         ) -> std::result::Result<tonic::Response<super::GameInfoBasic>, tonic::Status> {
             self.inner
                 .ready()
@@ -176,7 +170,7 @@ pub mod game_service_client {
         }
         pub async fn get_multiple_games_info(
             &mut self,
-            request: impl tonic::IntoRequest<super::MultipleGamesInfoRequest>,
+            request: impl tonic::IntoRequest<super::MultipleGamesRequest>,
         ) -> std::result::Result<
             tonic::Response<super::MultipleGamesInfoResponse>,
             tonic::Status,
@@ -201,7 +195,7 @@ pub mod game_service_client {
         }
         pub async fn get_game_detailed_info(
             &mut self,
-            request: impl tonic::IntoRequest<super::GameInfoRequest>,
+            request: impl tonic::IntoRequest<super::SingleGameRequest>,
         ) -> std::result::Result<
             tonic::Response<super::GameInfoDetailed>,
             tonic::Status,
@@ -270,7 +264,7 @@ pub mod game_service_client {
         }
         pub async fn delete_game(
             &mut self,
-            request: impl tonic::IntoRequest<super::DeleteGameRequest>,
+            request: impl tonic::IntoRequest<super::SingleGameRequest>,
         ) -> std::result::Result<
             tonic::Response<super::super::common::Empty>,
             tonic::Status,
@@ -304,18 +298,18 @@ pub mod game_service_server {
     pub trait GameService: Send + Sync + 'static {
         async fn get_single_game_info(
             &self,
-            request: tonic::Request<super::GameInfoRequest>,
+            request: tonic::Request<super::SingleGameRequest>,
         ) -> std::result::Result<tonic::Response<super::GameInfoBasic>, tonic::Status>;
         async fn get_multiple_games_info(
             &self,
-            request: tonic::Request<super::MultipleGamesInfoRequest>,
+            request: tonic::Request<super::MultipleGamesRequest>,
         ) -> std::result::Result<
             tonic::Response<super::MultipleGamesInfoResponse>,
             tonic::Status,
         >;
         async fn get_game_detailed_info(
             &self,
-            request: tonic::Request<super::GameInfoRequest>,
+            request: tonic::Request<super::SingleGameRequest>,
         ) -> std::result::Result<
             tonic::Response<super::GameInfoDetailed>,
             tonic::Status,
@@ -330,7 +324,7 @@ pub mod game_service_server {
         ) -> std::result::Result<tonic::Response<super::GameInfoBasic>, tonic::Status>;
         async fn delete_game(
             &self,
-            request: tonic::Request<super::DeleteGameRequest>,
+            request: tonic::Request<super::SingleGameRequest>,
         ) -> std::result::Result<
             tonic::Response<super::super::common::Empty>,
             tonic::Status,
@@ -420,7 +414,7 @@ pub mod game_service_server {
                     struct GetSingleGameInfoSvc<T: GameService>(pub Arc<T>);
                     impl<
                         T: GameService,
-                    > tonic::server::UnaryService<super::GameInfoRequest>
+                    > tonic::server::UnaryService<super::SingleGameRequest>
                     for GetSingleGameInfoSvc<T> {
                         type Response = super::GameInfoBasic;
                         type Future = BoxFuture<
@@ -429,7 +423,7 @@ pub mod game_service_server {
                         >;
                         fn call(
                             &mut self,
-                            request: tonic::Request<super::GameInfoRequest>,
+                            request: tonic::Request<super::SingleGameRequest>,
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
@@ -467,7 +461,7 @@ pub mod game_service_server {
                     struct GetMultipleGamesInfoSvc<T: GameService>(pub Arc<T>);
                     impl<
                         T: GameService,
-                    > tonic::server::UnaryService<super::MultipleGamesInfoRequest>
+                    > tonic::server::UnaryService<super::MultipleGamesRequest>
                     for GetMultipleGamesInfoSvc<T> {
                         type Response = super::MultipleGamesInfoResponse;
                         type Future = BoxFuture<
@@ -476,7 +470,7 @@ pub mod game_service_server {
                         >;
                         fn call(
                             &mut self,
-                            request: tonic::Request<super::MultipleGamesInfoRequest>,
+                            request: tonic::Request<super::MultipleGamesRequest>,
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
@@ -514,7 +508,7 @@ pub mod game_service_server {
                     struct GetGameDetailedInfoSvc<T: GameService>(pub Arc<T>);
                     impl<
                         T: GameService,
-                    > tonic::server::UnaryService<super::GameInfoRequest>
+                    > tonic::server::UnaryService<super::SingleGameRequest>
                     for GetGameDetailedInfoSvc<T> {
                         type Response = super::GameInfoDetailed;
                         type Future = BoxFuture<
@@ -523,7 +517,7 @@ pub mod game_service_server {
                         >;
                         fn call(
                             &mut self,
-                            request: tonic::Request<super::GameInfoRequest>,
+                            request: tonic::Request<super::SingleGameRequest>,
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
@@ -653,7 +647,7 @@ pub mod game_service_server {
                     struct DeleteGameSvc<T: GameService>(pub Arc<T>);
                     impl<
                         T: GameService,
-                    > tonic::server::UnaryService<super::DeleteGameRequest>
+                    > tonic::server::UnaryService<super::SingleGameRequest>
                     for DeleteGameSvc<T> {
                         type Response = super::super::common::Empty;
                         type Future = BoxFuture<
@@ -662,7 +656,7 @@ pub mod game_service_server {
                         >;
                         fn call(
                             &mut self,
-                            request: tonic::Request<super::DeleteGameRequest>,
+                            request: tonic::Request<super::SingleGameRequest>,
                         ) -> Self::Future {
                             let inner = Arc::clone(&self.0);
                             let fut = async move {
@@ -733,483 +727,5 @@ pub mod game_service_server {
     }
     impl<T: GameService> tonic::server::NamedService for GameServiceServer<T> {
         const NAME: &'static str = "game.GameService";
-    }
-}
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ReviewGameRequest {
-    #[prost(sfixed64, tag = "1")]
-    pub game_id: i64,
-    #[prost(int32, tag = "2")]
-    pub rating: i32,
-    #[prost(string, tag = "3")]
-    pub comments: ::prost::alloc::string::String,
-}
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct SingleGameRequest {
-    #[prost(sfixed64, tag = "1")]
-    pub game_id: i64,
-}
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GameReviews {
-    #[prost(message, repeated, tag = "1")]
-    pub reviews: ::prost::alloc::vec::Vec<GameReviewEntry>,
-}
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GameReviewEntry {
-    #[prost(sfixed64, tag = "1")]
-    pub user_id: i64,
-    #[prost(int32, tag = "2")]
-    pub rating: i32,
-    #[prost(string, tag = "3")]
-    pub comments: ::prost::alloc::string::String,
-    #[prost(int64, tag = "4")]
-    pub review_date_s: i64,
-}
-/// Generated client implementations.
-pub mod review_service_client {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    use tonic::codegen::http::Uri;
-    #[derive(Debug, Clone)]
-    pub struct ReviewServiceClient<T> {
-        inner: tonic::client::Grpc<T>,
-    }
-    impl ReviewServiceClient<tonic::transport::Channel> {
-        /// Attempt to create a new client by connecting to a given endpoint.
-        pub async fn connect<D>(dst: D) -> Result<Self, tonic::transport::Error>
-        where
-            D: TryInto<tonic::transport::Endpoint>,
-            D::Error: Into<StdError>,
-        {
-            let conn = tonic::transport::Endpoint::new(dst)?.connect().await?;
-            Ok(Self::new(conn))
-        }
-    }
-    impl<T> ReviewServiceClient<T>
-    where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
-        T::Error: Into<StdError>,
-        T::ResponseBody: Body<Data = Bytes> + Send + 'static,
-        <T::ResponseBody as Body>::Error: Into<StdError> + Send,
-    {
-        pub fn new(inner: T) -> Self {
-            let inner = tonic::client::Grpc::new(inner);
-            Self { inner }
-        }
-        pub fn with_origin(inner: T, origin: Uri) -> Self {
-            let inner = tonic::client::Grpc::with_origin(inner, origin);
-            Self { inner }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> ReviewServiceClient<InterceptedService<T, F>>
-        where
-            F: tonic::service::Interceptor,
-            T::ResponseBody: Default,
-            T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
-                >,
-            >,
-            <T as tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-            >>::Error: Into<StdError> + Send + Sync,
-        {
-            ReviewServiceClient::new(InterceptedService::new(inner, interceptor))
-        }
-        /// Compress requests with the given encoding.
-        ///
-        /// This requires the server to support it otherwise it might respond with an
-        /// error.
-        #[must_use]
-        pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.inner = self.inner.send_compressed(encoding);
-            self
-        }
-        /// Enable decompressing responses.
-        #[must_use]
-        pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.inner = self.inner.accept_compressed(encoding);
-            self
-        }
-        /// Limits the maximum size of a decoded message.
-        ///
-        /// Default: `4MB`
-        #[must_use]
-        pub fn max_decoding_message_size(mut self, limit: usize) -> Self {
-            self.inner = self.inner.max_decoding_message_size(limit);
-            self
-        }
-        /// Limits the maximum size of an encoded message.
-        ///
-        /// Default: `usize::MAX`
-        #[must_use]
-        pub fn max_encoding_message_size(mut self, limit: usize) -> Self {
-            self.inner = self.inner.max_encoding_message_size(limit);
-            self
-        }
-        pub async fn review_game(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ReviewGameRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::super::common::Empty>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/game.ReviewService/ReviewGame",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("game.ReviewService", "ReviewGame"));
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn delete_review(
-            &mut self,
-            request: impl tonic::IntoRequest<super::SingleGameRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::super::common::Empty>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/game.ReviewService/DeleteReview",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("game.ReviewService", "DeleteReview"));
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn get_game_reviews(
-            &mut self,
-            request: impl tonic::IntoRequest<super::SingleGameRequest>,
-        ) -> std::result::Result<tonic::Response<super::GameReviews>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/game.ReviewService/GetGameReviews",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("game.ReviewService", "GetGameReviews"));
-            self.inner.unary(req, path, codec).await
-        }
-    }
-}
-/// Generated server implementations.
-pub mod review_service_server {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    /// Generated trait containing gRPC methods that should be implemented for use with ReviewServiceServer.
-    #[async_trait]
-    pub trait ReviewService: Send + Sync + 'static {
-        async fn review_game(
-            &self,
-            request: tonic::Request<super::ReviewGameRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::super::common::Empty>,
-            tonic::Status,
-        >;
-        async fn delete_review(
-            &self,
-            request: tonic::Request<super::SingleGameRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::super::common::Empty>,
-            tonic::Status,
-        >;
-        async fn get_game_reviews(
-            &self,
-            request: tonic::Request<super::SingleGameRequest>,
-        ) -> std::result::Result<tonic::Response<super::GameReviews>, tonic::Status>;
-    }
-    #[derive(Debug)]
-    pub struct ReviewServiceServer<T: ReviewService> {
-        inner: _Inner<T>,
-        accept_compression_encodings: EnabledCompressionEncodings,
-        send_compression_encodings: EnabledCompressionEncodings,
-        max_decoding_message_size: Option<usize>,
-        max_encoding_message_size: Option<usize>,
-    }
-    struct _Inner<T>(Arc<T>);
-    impl<T: ReviewService> ReviewServiceServer<T> {
-        pub fn new(inner: T) -> Self {
-            Self::from_arc(Arc::new(inner))
-        }
-        pub fn from_arc(inner: Arc<T>) -> Self {
-            let inner = _Inner(inner);
-            Self {
-                inner,
-                accept_compression_encodings: Default::default(),
-                send_compression_encodings: Default::default(),
-                max_decoding_message_size: None,
-                max_encoding_message_size: None,
-            }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> InterceptedService<Self, F>
-        where
-            F: tonic::service::Interceptor,
-        {
-            InterceptedService::new(Self::new(inner), interceptor)
-        }
-        /// Enable decompressing requests with the given encoding.
-        #[must_use]
-        pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.accept_compression_encodings.enable(encoding);
-            self
-        }
-        /// Compress responses with the given encoding, if the client supports it.
-        #[must_use]
-        pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.send_compression_encodings.enable(encoding);
-            self
-        }
-        /// Limits the maximum size of a decoded message.
-        ///
-        /// Default: `4MB`
-        #[must_use]
-        pub fn max_decoding_message_size(mut self, limit: usize) -> Self {
-            self.max_decoding_message_size = Some(limit);
-            self
-        }
-        /// Limits the maximum size of an encoded message.
-        ///
-        /// Default: `usize::MAX`
-        #[must_use]
-        pub fn max_encoding_message_size(mut self, limit: usize) -> Self {
-            self.max_encoding_message_size = Some(limit);
-            self
-        }
-    }
-    impl<T, B> tonic::codegen::Service<http::Request<B>> for ReviewServiceServer<T>
-    where
-        T: ReviewService,
-        B: Body + Send + 'static,
-        B::Error: Into<StdError> + Send + 'static,
-    {
-        type Response = http::Response<tonic::body::BoxBody>;
-        type Error = std::convert::Infallible;
-        type Future = BoxFuture<Self::Response, Self::Error>;
-        fn poll_ready(
-            &mut self,
-            _cx: &mut Context<'_>,
-        ) -> Poll<std::result::Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-        fn call(&mut self, req: http::Request<B>) -> Self::Future {
-            let inner = self.inner.clone();
-            match req.uri().path() {
-                "/game.ReviewService/ReviewGame" => {
-                    #[allow(non_camel_case_types)]
-                    struct ReviewGameSvc<T: ReviewService>(pub Arc<T>);
-                    impl<
-                        T: ReviewService,
-                    > tonic::server::UnaryService<super::ReviewGameRequest>
-                    for ReviewGameSvc<T> {
-                        type Response = super::super::common::Empty;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ReviewGameRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ReviewService>::review_game(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ReviewGameSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/game.ReviewService/DeleteReview" => {
-                    #[allow(non_camel_case_types)]
-                    struct DeleteReviewSvc<T: ReviewService>(pub Arc<T>);
-                    impl<
-                        T: ReviewService,
-                    > tonic::server::UnaryService<super::SingleGameRequest>
-                    for DeleteReviewSvc<T> {
-                        type Response = super::super::common::Empty;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::SingleGameRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ReviewService>::delete_review(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = DeleteReviewSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/game.ReviewService/GetGameReviews" => {
-                    #[allow(non_camel_case_types)]
-                    struct GetGameReviewsSvc<T: ReviewService>(pub Arc<T>);
-                    impl<
-                        T: ReviewService,
-                    > tonic::server::UnaryService<super::SingleGameRequest>
-                    for GetGameReviewsSvc<T> {
-                        type Response = super::GameReviews;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::SingleGameRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ReviewService>::get_game_reviews(&inner, request)
-                                    .await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = GetGameReviewsSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                _ => {
-                    Box::pin(async move {
-                        Ok(
-                            http::Response::builder()
-                                .status(200)
-                                .header("grpc-status", "12")
-                                .header("content-type", "application/grpc")
-                                .body(empty_body())
-                                .unwrap(),
-                        )
-                    })
-                }
-            }
-        }
-    }
-    impl<T: ReviewService> Clone for ReviewServiceServer<T> {
-        fn clone(&self) -> Self {
-            let inner = self.inner.clone();
-            Self {
-                inner,
-                accept_compression_encodings: self.accept_compression_encodings,
-                send_compression_encodings: self.send_compression_encodings,
-                max_decoding_message_size: self.max_decoding_message_size,
-                max_encoding_message_size: self.max_encoding_message_size,
-            }
-        }
-    }
-    impl<T: ReviewService> Clone for _Inner<T> {
-        fn clone(&self) -> Self {
-            Self(Arc::clone(&self.0))
-        }
-    }
-    impl<T: std::fmt::Debug> std::fmt::Debug for _Inner<T> {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{:?}", self.0)
-        }
-    }
-    impl<T: ReviewService> tonic::server::NamedService for ReviewServiceServer<T> {
-        const NAME: &'static str = "game.ReviewService";
     }
 }
